Former-commit-id: abd1670fd575fcb9f8bf556bfb27b26779b5d056

+++ src/runtime/cuda.cpp
@@ -357,7 +357,7 @@ static CUfunction __get_kernel(void *user_context, CUmodule mod, const char* ent
 
 static size_t __buf_size(void *user_context, buffer_t* buf) {
     size_t size = 0;
-    for (int i = 0; i < sizeof(buf->stride) / sizeof(buf->stride[0]); i++) {
+    for (size_t i = 0; i < sizeof(buf->stride) / sizeof(buf->stride[0]); i++) {
         size_t total_dim_size = buf->elem_size * buf->extent[i] * buf->stride[i];
         if (total_dim_size > size)
             size = total_dim_size;
